fix(wp-i-pr2): clippy useless_conversion + catalogue u64 BER-OID helpers (pre-push rails)
Fix 1 (clippy): bindings/jvm/src/error.rs lines 210+218 — remove
`u64::from(*target_id)` identity conversions now that VTargetPackEmpty and
DuplicateTargetId carry `target_id: u64` (widened by REF-KLV-04). Changed
to plain `*target_id` dereference at both sites.

Fix 2 (no-public-usize ratchet): extend the BER-length allowlist pattern in
scripts/check/rust/no-public-usize.sh from `pub fn ber_(oid_)?len\b` to
`pub fn ber_(oid_)?len(_u64)?\b` so `ber_oid_len_u64(u64) -> usize` is
permitted on the same "encode-helper sizing, Rust-only" rationale as its
u32 sibling `ber_oid_len`. Kept return type as `usize` (consistent with
sibling; no public-api.txt change).

Fix 3 (surface-manifest ratchet): add three [[exempt]] entries to
tests/coverage/surface-manifest.toml for the new u64 BER-OID helpers
(ber_oid_len_u64, read_ber_oid_u64, write_ber_oid_u64), following the
existing bootstrap pattern of their u32 siblings.

No runtime-behavior change. C ABI unchanged. non_exhaustive stays 269.
public-api.txt diff is empty.

+++ bindings/jvm/src/error.rs
@@ -207,7 +207,7 @@ pub fn map_klv_encode_error(env: &mut JNIEnv, e: &KlvEncodeError) {
             throw_klv_encode(env, "RESERVED_TAG_IN_UNKNOWN", Some(u64::from(*tag)), &msg)
         }
         KlvEncodeError::VTargetPackEmpty { target_id } => {
-            throw_klv_encode(env, "VTARGET_PACK_EMPTY", Some(u64::from(*target_id)), &msg)
+            throw_klv_encode(env, "VTARGET_PACK_EMPTY", Some(*target_id), &msg)
         }
         KlvEncodeError::DuplicateTargetId { target_id } => {
             // Hoist the boxed tag so the `throw_klv_encode(env, "<CONST>", ...)`
@@ -215,7 +215,7 @@ pub fn map_klv_encode_error(env: &mut JNIEnv, e: &KlvEncodeError) {
             // error-mapping ratchet's per-constant grep (a brace-less arm with
             // this longer CONST would otherwise split the call across lines and
             // hide the constant from the grep).
-            let t = Some(u64::from(*target_id));
+            let t = Some(*target_id);
             throw_klv_encode(env, "DUPLICATE_TARGET_ID", t, &msg)
         }
         KlvEncodeError::ForbiddenStandaloneOffset { tag } => {
